+++ 24Hour.Data/IdentityModels.cs
@@ -35,7 +35,7 @@ namespace _24HourAPI.Models
         }
         public DbSet<Post> Posts { get; set; }
         public DbSet<Comment> Comments { get; set; }
-        public DbSet<Like> Likes { get; set; } //me..
+        public DbSet<Like> Likes { get; set; }
         public DbSet<Reply> Replies { get; set; }
         protected override void OnModelCreating(DbModelBuilder modelBuilder)
         {
