+++ 24Hour.Data/packages.config
@@ -1,16 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="EntityFramework" version="6.4.4" targetFramework="net48" />
+  <package id="EntityFramework" version="6.1.0" targetFramework="net48" />
   <package id="Microsoft.AspNet.Identity.Core" version="2.2.3" targetFramework="net48" />
   <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.3" targetFramework="net48" />
   <package id="Microsoft.AspNet.Identity.Owin" version="2.2.3" targetFramework="net48" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.7" targetFramework="net48" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.7" targetFramework="net48" />
   <package id="Microsoft.AspNet.WebApi.Owin" version="5.2.7" targetFramework="net48" />
-  <package id="Microsoft.Owin" version="4.2.0" targetFramework="net48" />
-  <package id="Microsoft.Owin.Security" version="4.2.0" targetFramework="net48" />
-  <package id="Microsoft.Owin.Security.Cookies" version="4.2.0" targetFramework="net48" />
-  <package id="Microsoft.Owin.Security.OAuth" version="4.2.0" targetFramework="net48" />
-  <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net48" />
+  <package id="Microsoft.Owin" version="3.0.1" targetFramework="net48" />
+  <package id="Microsoft.Owin.Security" version="3.0.1" targetFramework="net48" />
+  <package id="Microsoft.Owin.Security.Cookies" version="3.0.1" targetFramework="net48" />
+  <package id="Microsoft.Owin.Security.OAuth" version="3.0.1" targetFramework="net48" />
+  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net48" />
   <package id="Owin" version="1.0" targetFramework="net48" />
 </packages>
