--- 24Hour.Models/Likes/LikeCreate.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace _24Hour.Models.Likes
-{
-    public class LikeCreate
-    {
-        [Required]
-        public int PostId { get; set; }
-    }
-}
