--- 24Hour.Models/Likes/LikeListItem.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace _24Hour.Models.Likes
-{
-    public class LikeListItem
-    {
-        public int LikeId { get; set; }
-        public int PostId { get; set; }
-    }
-}
