--- 24HourAPI/Controllers/LikeController.cs
@@ -1,35 +1,0 @@
-﻿using _24Hour.Models.Likes;
-using _24Hour.Services;
-using Microsoft.AspNet.Identity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
-using System.Web.Http;
-
-namespace _24HourAPI.Controllers
-{
-    [Authorize]
-    public class LikeController : ApiController
-    {
-       private LikeService CreateLikeService()
-        {
-            var userId = Guid.Parse(User.Identity.GetUserId());
-            var noteService = new LikeService(userId);
-            return noteService;
-        }
-        public IHttpActionResult Post(LikeCreate like)
-        {
-            if (!ModelState.IsValid)
-                return BadRequest(ModelState);
-
-            var service = CreateLikeService();
-
-            if (!service.CreateLike(like))
-                return InternalServerError();
-
-            return Ok();
-        }
-    }
-}
